requirements.txt
flask
python-dotenv
line-bot-sdk==3.*
ckiptagger[tf,gdown]==0.2.1
tensorflow==2.13.1
